+++ mobile_backend/src/main/java/com/example/mobile/controller/ProfileController.java
@@ -2,11 +2,7 @@ package com.example.mobile.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import com.example.mobile.dto.ProfileDTO;
 import com.example.mobile.dto.ResponseDTO;
@@ -16,7 +12,7 @@ import com.example.mobile.repository.UserRepository;
 import com.example.mobile.service.AuthStaticService;
 import com.example.mobile.service.UserService;
 
-import io.swagger.v3.oas.annotations.parameters.RequestBody;
+
 
 @RestController
 @RequestMapping("/profile")
