+++ mobile_frontend/app/src/main/java/com/example/myapplication/convert/ImageConvertProfile.java
@@ -1,0 +1,22 @@
+package com.example.myapplication.convert;
+
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.util.Base64;
+import android.widget.ImageView;
+
+import java.io.ByteArrayOutputStream;
+
+public class ImageConvertProfile {
+    public static String bitMapToBase64(Bitmap bitmap) {
+        ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
+        bitmap.compress(Bitmap.CompressFormat.PNG, 100, byteArrayOutputStream);
+        byte[] byteArray = byteArrayOutputStream.toByteArray();
+        return Base64.encodeToString(byteArray, Base64.NO_WRAP); // Use NO_WRAP to avoid line breaks
+    }
+
+    public static Bitmap base64ToBitMap(String encodedString) {
+        byte[] decodedString = Base64.decode(encodedString, Base64.DEFAULT);
+        return BitmapFactory.decodeByteArray(decodedString, 0, decodedString.length);
+    }
+}
