+++ mobile_frontend/app/src/main/java/com/example/myapplication/ui/adapters/ProfileAdapter.java
@@ -13,6 +13,7 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.myapplication.R;
 import com.example.myapplication.convert.ImageConvert;
+import com.example.myapplication.convert.ImageConvertProfile;
 import com.example.myapplication.network.model.dto.FriendViewDTO;
 import com.example.myapplication.ui.activities.ProfileActivity;
 
@@ -56,7 +57,7 @@ public class ProfileAdapter extends RecyclerView.Adapter<ProfileAdapter.ViewHold
 
 //         Hiển thị hình ảnh trong ImageView
         ImageView imageView = holder.avatarImageView;
-        imageView.setImageBitmap(ImageConvert.base64ToBitMap(friend.getAvatar()));
+        imageView.setImageBitmap(ImageConvertProfile.base64ToBitMap(friend.getAvatar()));
 
         holder.itemView.setOnClickListener(new View.OnClickListener() {
             @Override
